move to sofa/plugins
