+++ Sofa/CMakeLists.txt
@@ -16,3 +16,6 @@ add_subdirectory(framework)
 
 ### Component
 add_subdirectory(Component)
+
+### Plugins
+add_subdirectory(plugins)
