+++ Sofa/plugins/CMakeLists.txt
@@ -1,0 +1,6 @@
+cmake_minimum_required(VERSION 3.12)
+
+find_package(Sofa.Config) # contains sofa cmake macros
+
+sofa_add_plugin(SceneChecking SceneChecking ON)
+
