+++ Sofa/plugins/SceneChecking/SceneCheckingConfig.cmake.in
