+++ Sofa/plugins/SceneChecking/src/SceneChecking/SceneCheckDuplicatedName.h
@@ -22,41 +22,34 @@
 #pragma once
 
 #include <SceneChecking/config.h>
+#include <SceneChecking/SceneCheck.h>
 
-#include <sofa/simulation/SceneLoaderFactory.h>
-#include <sofa/simulation/Visitor.h>
+#include <map>
+#include <sstream>
 
-#include <SceneChecking/SceneCheckerVisitor.h>
-using scenechecking::SceneCheckerVisitor;
-
-
-namespace _scenechecking_
+namespace sofa::_scenechecking_
 {
-
-/// to be able to react when a scene is loaded
-class SOFA_SCENECHECKING_API SceneCheckerListener : public sofa::simulation::SceneLoader::Listener
+    
+class SOFA_SCENECHECKING_API SceneCheckDuplicatedName : public SceneCheck
 {
 public:
-    static SceneCheckerListener* getInstance();
-    virtual ~SceneCheckerListener() {}
-
-    virtual void rightAfterLoadingScene(sofa::simulation::NodeSPtr node) override;
-
-    // Do nothing on reload
-    virtual void rightBeforeReloadingScene() override {}
-    virtual void rightAfterReloadingScene(sofa::simulation::NodeSPtr node) override
-    {
-        SOFA_UNUSED(node);
-    }
+    virtual ~SceneCheckDuplicatedName() {}
+    typedef std::shared_ptr<SceneCheckDuplicatedName> SPtr;
+    static SPtr newSPtr() { return SPtr(new SceneCheckDuplicatedName()); }
+    virtual const std::string getName() override;
+    virtual const std::string getDesc() override;
+    void doInit(sofa::simulation::Node* node) override;
+    void doCheckOn(sofa::simulation::Node* node) override;
+    void doPrintSummary() override;
 
 private:
-    SceneCheckerListener();
-    SceneCheckerVisitor m_sceneChecker;
+    bool m_hasDuplicates;
+    std::stringstream m_duplicatedMsg;
 };
 
-} // namespace _scenechecking_
+} // namespace sofa::_scenechecking_
 
-namespace scenechecking
+namespace sofa::scenechecking
 {
-using _scenechecking_::SceneCheckerListener;
-} // namespace scenechecking
+    using _scenechecking_::SceneCheckDuplicatedName;
+}
