+++ applications/plugins/CMakeLists.txt
@@ -8,8 +8,6 @@ if(SOFA_BUILD_TESTS OR SOFA_BUILD_RELEASE_PACKAGE)
     add_subdirectory(SofaTest)
 endif()
 
-sofa_add_plugin(SceneChecking SceneChecking ON)
-
 sofa_add_plugin(CollisionOBBCapsule CollisionOBBCapsule)
 
 sofa_add_subdirectory_external(SofaHighOrder SofaHighOrder)
